code/WashWise/requirements.txt
roboflow-inference
opencv-python
python-dotenv
numpy